Add simple test for UserService
I reused the logic in the todo method but also return a User object. I
assume that this is what we want from the UserService since it is
located in a domain package.

Next up is refactoring the code to rid the domain from infrastructure details a la Ports and Adapters.

See: https://en.wikipedia.org/wiki/Hexagonal_architecture_(software)
Signed-off-by: Eric Thelin <extern.eric.thelin@digg.se>