feat: pre-emptively install node packages
All-though our MegaLinter configuration will install node packages prior
to running any linting, here we install node packages preemptively to
avoid issues when MegaLinter tries to install the node packages inside
its Docker container when behind a corporate proxy. This is not
supported by the current MegaLinter setup.

With this extra step, the node modules will already be present when
MegaLinter tries to install them and the install will be skipped instead
of causing an error.

Note: For the node package install to work here it is imperative that
the machine running the code quality script is properly configured for
running behind the proxy.

Signed-off-by: Eric Thelin <extern.eric.thelin@digg.se>